Move CompressedPoV to primitives.
