+++ primitives/Cargo.toml
@@ -26,6 +26,9 @@ bitvec = { version = "0.20.1", default-features = false, features = ["alloc"] }
 frame-system = { git = "https://github.com/paritytech/substrate", branch = "master", default-features = false }
 hex-literal = "0.3.1"
 parity-util-mem = { version = "0.9.0", default-features = false, optional = true }
+thiserror = "1.0.23"
+[target.'cfg(not(target_os = "unknown"))'.dependencies]
+zstd = "0.5.0"
 
 [dev-dependencies]
 sp-serializer = { git = "https://github.com/paritytech/substrate", branch = "master" }
